Fix sign up errors
Two errors occured while sign up, the activation link was incorrect
and the JSON containing the password value was incorrect.

The first problem was solved by getting the host and port separately
to build the URL for the activation link.

The second problem was solved by removing a trailing space before
the password value.