Fix failing test on Java 11
Mockito can't for some reason not mock the ContainerResponseContext
interface in Java 11 (works in Java 12).

The solution is to add a dummy implementation of the class, together
with a "count" method so that we can verify that the getHeaders
method was called. Unfortunately the API for the
ContainerResponseContext does not have an "addHeader" and instead
does "getHeaders", which can't return a copy of the headers so that
we can verify it's content.

This means that we can't verify that the correct key/value is added
to the headers map, but at least we know that the method getHeaders
wss called.