Refactor AccountResource
The account resource class had too many dependencies and did things
such as sending email, making http requests and encode/decode data.

This made the class difficult to test and the tests where not very
readable.

Two new classes are added, account and email client.The account is
used by account resource and the email client is used by account.

The flow is still the same, but the classes are now smaller and it
lets us mock more data.