Add DashboardResource integration tests
Added a test for dashboard resource. The test uses a reg-ex, since the data
in the database differs from time to time.

This means that all data is verified, except for the session data. It only
checks that there is a sessions key and that there exists an list is present.
If this list is empty or not, or how the session data looks like is not
verified.

Since the test needs both the report service as well as a database
connection, the test is disabled.
The test is verified locally with the report service and database
connection.