Add unauthorized dashboard test
Added a test that verifies that an unauthorixed request can't
access dashboard data.

The TestHttpServletRequest class needed to be updated to that it
won't return a new http session each time.

The TestHttpSession needed to be updated so that the key/value
is saved when setAttribute is called and that the value is
returned when getAttribute is called.